firmware: payload: Add a test case to trigger PMP access failure
This adds a payload test case to trigger PMP check failure.

It sets up 2GiB virtual to physical address translation, but after
satp is written with sv39 on, immediately the core raises a PMP
access failure.

Disabling PMP region 0 makes this issue disappear.
PMP0: 0x0000000080000000-0x000000008001ffff (A)

This case is not reproducible on QEMU, but on real hardware.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>